+++ Sources/Sentry/SentryTracer.m
@@ -588,16 +588,22 @@ static NSLock *profilerLock;
                                    description:type];
     [appStartSpan setStartTimestamp:appStartMeasurement.appStartTimestamp];
 
+    SentrySpan *premainSpan = [self buildSpan:appStartSpan.context.spanId
+                                    operation:operation
+                                  description:@"Pre Runtime Init"];
+    [premainSpan setStartTimestamp:appStartMeasurement.appStartTimestamp];
+    [premainSpan setTimestamp:appStartMeasurement.runtimeInitTimestamp];
+
     SentrySpan *runtimeInitSpan = [self buildSpan:appStartSpan.context.spanId
                                         operation:operation
-                                      description:@"Pre main"];
-    [runtimeInitSpan setStartTimestamp:appStartMeasurement.appStartTimestamp];
-    [runtimeInitSpan setTimestamp:appStartMeasurement.runtimeInitTimestamp];
+                                      description:@"Runtime Init to Pre Main Initializers"];
+    [runtimeInitSpan setStartTimestamp:appStartMeasurement.runtimeInitTimestamp];
+    [runtimeInitSpan setTimestamp:appStartMeasurement.moduleInitializationTimestamp];
 
     SentrySpan *appInitSpan = [self buildSpan:appStartSpan.context.spanId
                                     operation:operation
                                   description:@"UIKit and Application Init"];
-    [appInitSpan setStartTimestamp:appStartMeasurement.runtimeInitTimestamp];
+    [appInitSpan setStartTimestamp:appStartMeasurement.moduleInitializationTimestamp];
     [appInitSpan setTimestamp:appStartMeasurement.didFinishLaunchingTimestamp];
 
     SentrySpan *frameRenderSpan = [self buildSpan:appStartSpan.context.spanId
@@ -608,7 +614,7 @@ static NSLock *profilerLock;
 
     [appStartSpan setTimestamp:appStartEndTimestamp];
 
-    return @[ appStartSpan, runtimeInitSpan, appInitSpan, frameRenderSpan ];
+    return @[ appStartSpan, premainSpan, runtimeInitSpan, appInitSpan, frameRenderSpan ];
 }
 
 - (void)addMeasurements:(SentryTransaction *)transaction