+++ Sources/Sentry/SentryTransportAdapter.m
@@ -58,10 +58,8 @@ SentryTransportAdapter ()
                                                                attachments:attachments];
     [items addObjectsFromArray:additionalEnvelopeItems];
 
-    SentryEnvelopeHeader *envelopeHeader =
-        [[SentryEnvelopeHeader alloc] initWithId:event.eventId
-                                         sdkInfo:self.options.sdkInfo
-                                    traceContext:traceContext];
+    SentryEnvelopeHeader *envelopeHeader = [[SentryEnvelopeHeader alloc] initWithId:event.eventId
+                                                                       traceContext:traceContext];
     SentryEnvelope *envelope = [[SentryEnvelope alloc] initWithHeader:envelopeHeader items:items];
 
     [self sendEnvelope:envelope];
@@ -76,10 +74,8 @@ SentryTransportAdapter ()
                                                                attachments:attachments];
     [items addObject:[[SentryEnvelopeItem alloc] initWithSession:session]];
 
-    SentryEnvelopeHeader *envelopeHeader =
-        [[SentryEnvelopeHeader alloc] initWithId:event.eventId
-                                         sdkInfo:self.options.sdkInfo
-                                    traceContext:traceContext];
+    SentryEnvelopeHeader *envelopeHeader = [[SentryEnvelopeHeader alloc] initWithId:event.eventId
+                                                                       traceContext:traceContext];
 
     SentryEnvelope *envelope = [[SentryEnvelope alloc] initWithHeader:envelopeHeader items:items];
 
@@ -90,9 +86,7 @@ SentryTransportAdapter ()
 {
     SentryEnvelopeItem *item = [[SentryEnvelopeItem alloc] initWithUserFeedback:userFeedback];
     SentryEnvelopeHeader *envelopeHeader =
-        [[SentryEnvelopeHeader alloc] initWithId:userFeedback.eventId
-                                         sdkInfo:self.options.sdkInfo
-                                    traceContext:nil];
+        [[SentryEnvelopeHeader alloc] initWithId:userFeedback.eventId traceContext:nil];
     SentryEnvelope *envelope = [[SentryEnvelope alloc] initWithHeader:envelopeHeader
                                                            singleItem:item];
     [self sendEnvelope:envelope];