+++ CHANGELOG.md
@@ -5,12 +5,14 @@
 ### Features
 
 - Enhance the UIViewController breadcrumbs with more data (#1945)
+- feat: Add extra app start span (#1952)
 - Add enableAutoBreadcrumbTracking option (#1958)
 
 ### Fixes
 
 - Properly sanitize the event context and SDK information (#1943)
 - Don't send error 429 as `network_error` (#1957)
+- Deprecate not needed option `sdkInfo` (#1960)
 
 ## 7.20.0
 
@@ -41,7 +43,7 @@
 ## 7.18.0
 
 ### Features
- 
+
 - Replace tracestate header with baggage (#1867)
 
 ### Fixes